Move file accessor interfaces and NIO implementations to :files
