+++ platforms/core-runtime/files/src/test/groovy/org/gradle/internal/file/PosixFilePermissionConverterTest.groovy
@@ -14,13 +14,20 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.nativeintegration.filesystem.jdk7
+package org.gradle.internal.file
 
+import org.gradle.internal.file.nio.PosixFilePermissionConverter
 import spock.lang.Specification
 
 import java.nio.file.attribute.PosixFilePermission
 
-import static java.nio.file.attribute.PosixFilePermission.*
+import static java.nio.file.attribute.PosixFilePermission.GROUP_EXECUTE
+import static java.nio.file.attribute.PosixFilePermission.GROUP_READ
+import static java.nio.file.attribute.PosixFilePermission.GROUP_WRITE
+import static java.nio.file.attribute.PosixFilePermission.OTHERS_READ
+import static java.nio.file.attribute.PosixFilePermission.OWNER_EXECUTE
+import static java.nio.file.attribute.PosixFilePermission.OWNER_READ
+import static java.nio.file.attribute.PosixFilePermission.OWNER_WRITE
 
 class PosixFilePermissionConverterTest extends Specification {
     def "converts Set<PosixFilePermission to int representation"() {
