+++ platforms/core-runtime/files/src/test/groovy/org/gradle/internal/file/nio/PosixJdk7FilePermissionHandlerTest.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.nativeintegration.filesystem.jdk7
+package org.gradle.internal.file.nio
 
 import org.gradle.test.fixtures.file.TestNameTestDirectoryProvider
 import org.gradle.test.precondition.Requires
