+++ platforms/core-runtime/native/src/main/java/org/gradle/internal/nativeintegration/filesystem/services/GenericFileSystem.java
@@ -23,9 +23,9 @@ import org.gradle.internal.SystemProperties;
 import org.gradle.internal.file.FileException;
 import org.gradle.internal.file.FileMetadata;
 import org.gradle.internal.file.FileMetadataAccessor;
+import org.gradle.internal.file.FileModeAccessor;
+import org.gradle.internal.file.FileModeMutator;
 import org.gradle.internal.file.StatStatistics;
-import org.gradle.internal.nativeintegration.filesystem.FileModeAccessor;
-import org.gradle.internal.nativeintegration.filesystem.FileModeMutator;
 import org.gradle.internal.nativeintegration.filesystem.FileSystem;
 import org.gradle.internal.nativeintegration.filesystem.Symlink;
 import org.slf4j.Logger;
