+++ platforms/core-runtime/native/src/test/groovy/org/gradle/internal/nativeintegration/filesystem/services/NativePlatformBackedFileMetadataAccessorTest.groovy
@@ -18,6 +18,7 @@ package org.gradle.internal.nativeintegration.filesystem.services
 
 import net.rubygrapefruit.platform.Native
 import net.rubygrapefruit.platform.file.Files
+import org.gradle.internal.file.AbstractFileMetadataAccessorTest
 import org.gradle.internal.file.FileMetadata
 import org.gradle.internal.file.FileMetadataAccessor
 import org.gradle.util.UsesNativeServices
